Travis: set git clone depth to 1
By default Travis CI clones repositories to a depth of 50 commits.

With this flag Jetpack will be cloned only to depth 1.

Note that if you use a depth of 1 and have a queue of jobs, Travis CI won’t build commits that are in the queue when you push a new commit.

https://docs.travis-ci.com/user/customizing-the-build/#Git-Clone-Depth