Markdown block: fixes caret in live preview
After formatting the source introduced by the user, React re-renders the
component and this makes the caret to move to the beginning of the
live-preview component's content. We need to save the caret position as
the user enters new content, and then restore it when the component has
ben re-rendered.