Markdown block live preview: new lines issue fixed
The user had to press the enter key twice to be able to add a new line
after a heading. This commit makes the live preview component to check
if there was a change in the formatted markdown. When the user adds an
empty line, the markdown parser does not add a new paragraph because the
line is empty, and therefore there's no need to rerender the component
and the new line entered by the user is not removed.