Infinite Scroll: Change retrieval method for privacy policy link
As `the_privacy_policy_link` displays a pre-formatted link, it was not being incorporated into
the `blog-credits` div. The 'Privacy Policy' link text was also not easily accessible for
translation. Using `get_privacy_policy_url()` allowed for the formatting to match the rest of
the credits on the footer, and this secured the correct behavior.