Documentation update, bug fix for user checkout and enhancement.
Updated the documentation to better walk the hosts through the setup and expectations for the plugin.

There was a bug when checking out as a user and having the plan automatically provisioned. A new row was being inserted to store the next url and details instead of updating the existing row for the product order.  This is now an update instead of an insert.

Correctly handling situations where the provision endpoint returns true but the domain does not resolve. This will now store this in the jetpack_provisioning_details field to inform the host in place of the next_url