update make_api_request to provide more details to logActivity
The make_api_request function previously provided an error to the user that was not ideal for tracking down issues and required enabling Module Logging to catch the Exception and let the host know what went wrong.

If a siteurl is present on the request  the function now returns an error string with more information. WHMCS will append this to the activity log with a link to the Service Id it was placed from.

Also updated the remaining calls of logModuleCall to store a trace for outstanding Exceptions