Added line breaks to namespace attribute.
Before the geolocation tool would add the namespace attribute that was not separated by anything in the beginning of the string. This caused issues like #9885, where a plugin would add their own namespacing and cause a validation error because Jetpack would add a namespace right next to it without any spacing.

This PR, however, does not fix the problem of duplicate namespace attributes, which would also cause validation errors. In case of duplicate namespaces we need to disable Jetpack's functionality one plugin at a time by adding them to conflicts.