Contact Form: Personal Data Eraser Tests
https://github.com/Automattic/jetpack/pull/9998 addressed some pagination
bugs in the Personal Data Eraser for Jetpack's Contact Form. The best way
to test at the time was to manually edit the `$per_page` variable, then
submit some manual eraser requests.

Refactor the paginating function to accept a `$per_page` argument for use
by the tests. Don't expose that parameter to the "public" eraser function
in case core decides to start passing a third parameter in the future.

Add a pagination test.