Updates to make_api_request, addeded error handling for provisioning and cancellation and code cleanup
Updated the make_api_request to simply make the api request and no longer do any processing of the response. If there is a response and no curl errors the response will be json_decoded and now sent to the function that made the request.

Added 2 functions for handling errors specific to provisioning and cancelling a jetpack plan. These are intended to be able to parse the response from a request to provision or cancel including looking at the status code and the message in the response body if there is one to provide a useful error message.

Also updated the docs for most functions as they were either incomplete/incorrect or missing details about params and expectation.