Contact Form: Improve Email Address Handling Personal Data Eraser (#10088)
* Contact Form: Improve Email Address Handling Personal Data Eraser

We need to alter core's search query to narrow the search results.
Previously, we did so by searching for the email address, extracting that
address from the search query, and then using that extracted email address
to build a more refined query.

However, we weren't taking into account that the string representing the
email address can change when used in an SQL query. Most significantly, if
the email address has an underscore in it, that character will be LIKE
escaped: slashes will be added. Then it will be SQL escaped: more slashes.

After extracting the address, we were then LIKE and SQL escaping it
*again* in for use in our new, more refined, query.

The result: if an email address has an underscore in it, the Personal Data
Exporter/Eraser didn't work at all.

Instead, let's just hang on to the email address in an object property and
use it directly when we need to. The search parameter is now just a
special flag we use for determining which queries need to be filtered/
refined.

* Contact Form: Personal Data Eraser Tests

https://github.com/Automattic/jetpack/pull/9998 addressed some pagination
bugs in the Personal Data Eraser for Jetpack's Contact Form. The best way
to test at the time was to manually edit the `$per_page` variable, then
submit some manual eraser requests.

Refactor the paginating function to accept a `$per_page` argument for use
by the tests. Don't expose that parameter to the "public" eraser function
in case core decides to start passing a third parameter in the future.

Add a pagination test.