PHPCS: remove superfluous PHPCS annotations
The annotation in the `modules/custom-css/csstidy/class.csstidy_print.php` file wasn't ignoring anything.
The annotation in the `modules/custom-css/csstidy/class.csstidy.php` file was hiding some very real issues with the code.

While I understand that the code is sourced elsewhere and should be fixed at the source and this file synced after that, shipping the code as is without applying the suggested fixes is not a good idea either.