Clean up duplicate post types in featured content
Add array_unique for the post_types list.

array_merge appends arrays, and since post is set as the default, it means that for many post will be listed twice.

For example in Canard (an Automattic theme) where post_type is set to post and page the post_type will become array( 'post', 'post', 'page' )
https://github.com/Automattic/themes/blob/b8444897830573b47505a19623e1f3459a45aac4/canard/inc/jetpack.php#L22

By filtering the list we prevent the register_taxonomy_for_object_type from being called multiple times for the same post type, and we ensure the post_types are listed once when we later call get_posts.

To be honest I don't imagine it will make a huge amount of difference to anything but I was debugging a featured content issue in my theme and saw this and it confused me. This eliminates that confusion.