CSS concatenation: fix differences between what the 2 lists.
One list is for the files being dequeued.
The other list is for the content of those files to be concatenated into jetpack.css.

The 2 lists should match, but that was not the case. I updated them.

Noting that I also removed The IDC file since it is meant to be used in wp-admin only.