update math_form() method input tag
Adds a single space before each line break in the jetpack_protect_num input tag in order to allow a space to be maintained between the tag name and the attributes when output is buffered. This would maintain compatibility with front end login plugins that need to buffer the output and collect it in ob_get_contents() so it can be displayed in a widget or main content.