Add param to allow customizing boosting of fields searched
This adds a param which can for now only be modified via filters which provide more control over how much importance we place on each field that is searched. Currently the following field boost values can be adjusted via this new `excess_boost` param:

* `title`
* `tag.name`
* `category.name`
* `author_login`
* `author`

The following fields are searched however they can not be boosted when terms are found in them:

* `content`
* `excerpt`

The boosting given is in excess of preconfigured boost and are multiplied to it thus a positive number must be given.

Caution: We may at some point adjust the default baseline boosting to provide better results which may amplify any manual fine tuning done with this filter.