Unlaunched sites: Set a blog option of "unlaunched" to sites that are created private.
This will enable us to distinguish sites that are created private from sites that are deliberately private, so that we can show a coming soon page on them and a pre-launch banner.

This diff does several things:

It sets a blog option of launch-status to unlaunched when users create a site which is private.
It enables users to publish posts and pages to sites which have the launch-status blog option, even if their email address is unverified (this is OK because sites which are  unlaunched aren't public, so there's no spam concerns, and they can't update the status to be launched without verifying their email address)
It returns the launch-status of a site in the API
Moves the functionality into a lib so we can use it in multiple places

Differential Revision: D19586

This commit syncs r182320-wpcom.