General: Add badge if API is sandboxed (#10373)
#### Changes proposed in this Pull Request:

This adds a very basic implementation that displays a badge if the `JETPACK__SANDBOX_DOMAIN` constant is in use, hence the API is sandboxed.  I don't think we need translation here. Only a12s will be using this constant anyway. 

It doesn't actually check the health of the sandboxed connection, only if a different URL is being used. 

![screen shot 2018-10-22 at 7 58 29 pm](https://user-images.githubusercontent.com/7129409/47326139-5c014d00-d635-11e8-88b7-acd9a83e9af2.png)

#### Testing instructions:

<!--
Add as many details as possible to help others reproduce the issue and test the fix.
"Before / After" screenshots can also be very helpful when the change is visual.

Would you like this feature to be tested by Beta testers as well?
Please add instructions to to-test.md in a new commit as part of your PR.
-->

*Set `JETPACK__SANDBOX_DOMAIN` to point to your sandbox URL in wp-config.php 
*You should see the badge in all views of the main Jetpack admin 
*You should not expect to see it in any of the headers that are not rendered in React

*Clear the constant - the badge should be gone. 

<!-- Add the following only if this is meant to be in changelog -->
#### Proposed changelog entry for your changes:
I think none needed.