Update is_vip() to return the value of the WPCOM_IS_VIP_ENV constant if it exists or false.
Previously, the `is_vip()` function returned `false` without checking anything. This was causing VIP Go sites to not be caught by the new sites filter in Calypso's Gutenberg route. See D20666-code for more details.

`wpcom_is_vip()` is deprecated on VIP Go and the `WPCOM_IS_VIP_ENV` constant is now the recommended way of checking if the if the current site is VIP.