Photon: Do not image_downsize for REST API requests in the edit context (#10592)
Photon: Do image_downsize for REST API requests on the wp/v2/media endpoint in the edit context.

For tests, we need to clean the `Jetpack_Photon` state before each test, which is annoying because of the singleton/static class property. Unfortunately, the tests are not compatible with PHP 5.3, though the feature is.