Contact form: Add jetpack-contact-form-shortcode-preview for styling … (#10684)
…in the editor

<!--- Provide a general summary of your changes in the Title above -->

Fixes broken styling in the Gutenberg editor together with the following PR 
https://github.com/Automattic/wp-calypso/pull/28735

#### Changes proposed in this Pull Request:
<!--- Explain what functional changes your PR includes --> 
* Adds a class so that we can better target the styles. 
In the classic edior the styles are loaded in a iframe so they can be more generic. 
in the gutenberg editor they need to be more specific so that we don't conflict with the existing plugins. 

#### Testing instructions:
Follow the testing instructions found in https://github.com/Automattic/wp-calypso/pull/28735