Improve the matching of the youtube shortcode reversal again.
Merges r184358-wpcom

In D17568-code and D17868-code we improved this code.
Those changes accidentally got reversed in D18511-code
I was reviewing #10224 which would merge the original
changes into Jetpack and noted the 's' modifier.
Unfortunately we can't include that here because it means on long but malformed post content we go into an
infinite loop because of the .*? in the RegEx.
This means we can't match objects/mangled objects which have a newline within them but we can match more
that we were

Diff: D21652-code