Jetpack Affiliate: add affiliate code to JITM links if it exists.
Introduce method add_code_as_query_arg In Jetpack_Affiliate class to fetch and add the affiliate code in a single call and avoid repeating it. Refactor code in Jetpack class to use this method instead of inline code. Added new unit test to check the correct behavior of the new method