Add review to the list of allowed comment types for the /comments/$comment_id (#10855)
#### Changes proposed in this Pull Request:

See p90Yrv-Tz-p2.

Prior to WooCommerce 3.5, product reviews were stored as comments with an empty comment type (''). After 3.5, these are now marked with the review type.

A few things relied on being able to update comment status using the WordPress.com REST API, such as the mobile apps and web notifications.

This PR whitelists the review type to restore that behavior. 

This was originally part of #10786 -- but separated out since it is a separate fix for the mobile apps.

#### Testing instructions:
* Apply this PR/change.
* Make sure you have WooCommerce with a product setup. Create a product review notification. Take note of the ID.
* Go to https://developer.wordpress.com/docs/api/console/ and point towards your test Jetpack site.
* Make a request to GET /sites/:site/comments/:comment (find via the dropdown) and verify that it does not 404.

#### Proposed changelog entry for your changes:
N/A