Search: Re-use should_handle_query filter elsewhere in class.
Seemed to be checking the same condition multiple places
(`$query->is_main_query() && $query->is_search()`) but only wrapping it
in a filter on the main `filter__posts_pre_query()`. The upshot of this
means that if someone did want to hook into that and use elasticsearch
outside of a main query or search they can't as it will get stopped
elsewhere by other places not using the filter. Thus rendering the
filter almost pointless (AFAICT).

By re-using the filter in other places that this check is done, the end
user is free to specify that they want to use jetpack search for the
query even if it's not a search query or the main query.