Remove some offending formatting-related ESLint rules
Removed some rules that are:
- already configured in `eslint-config-wpcalypso`
- disabled in `eslint-config-prettier`
- enabled again in `.eslintrc.js`, causing unwanted lint errors

One special case is `quote-props` with `keywords: true`. Prettier doesn't quote reserved
keywords used as property names, and the quoting is only needed for ES3 compatibility.
The Babel plugin `transform-es3-member-expression-literals` should take care of that: in
shipped code, the properties should be quoted again.