Connection banner: display banner inside React page.
Fixes #11071

- The fullscreen connnection banner appearing on Jetpack > Dashboard previously overlaid all other
information on the page. This included useful information like footer links, but also potential error
messages.
- This PR adds a new empty div to the main react admin page structure when a site is not connected to
WordPress.com, and then moves the content from the notice that contains the connection banner to inside
that empty div, in the middle of the page.

This creates a difference between the fullscreen connection banner in the plugins menu and the one in
Jetpack > Dashboard: since the Jetpack dashboard already includes a masthead with a Jetpaack logo, I
removed the logo from the connection banner there.