Search: Re-use should_handle_query filter elsewhere in class. (#10990)
<!--- Provide a general summary of your changes in the Title above -->

Fixes N/A (?)

#### Changes proposed in this Pull Request:
<!--- Explain what functional changes your PR includes -->

Seemed to be checking the same condition multiple places
(`$query->is_main_query() && $query->is_search()`) but only wrapping it
in a filter on the main `filter__posts_pre_query()`. The upshot of this
means that if someone did want to hook into that and use elasticsearch
outside of a main query or search they can't as it will get stopped
elsewhere by other places not using the filter. Thus rendering the
filter almost pointless (AFAICT).

By re-using the filter in other places that this check is done, the end
user is free to specify that they want to use jetpack search for the
query even if it's not a search query or the main query.

#### Testing instructions:
<!-- Please include detailed testing steps, explaining how to test your change. -->
<!-- Bear in mind that context you working on is not obvious for everyone.  -->
<!-- Adding "simple" configuration steps will help reviewers to get to your PR as quickly as possible. -->
<!-- "Before / After" screenshots can also be very helpful when the change is visual. -->

* Attempt to override default behaviour of search by using filter `jetpack_search_should_handle_query` observe that query doesn't actually send to Elasticsearch even though you are using the filter.

* Switch to this branch, now do the same and observe (through query monitor for example) that query is indeed being sent to Elasticsearch.

#### Proposed changelog entry for your changes:
<!-- Please do not leave this empty. If no changelog entry needed, state as such. -->

* Extend `jetpack_search_should_handle_query` filter so that you are able to use jetpack search if not a main query or not a search query.