Subscriptions: display the checkboxes for logged in users too.
In #9983, we changed the way the subscription fields were hooked into the comment form.
The comment_form_after_fields action we started using is not triggered for logged in users.

comment_form_submit_button seems to be the only better filter right now.
I also added some CSS so the fields are never too close of the submit button.