Apply fixed position to carousel overlay
The overlay's container is fixed, which "sticks" to the top of the
viewport. If the overlay is absolute instead of fixed and the content
overflows, this may allow for scrolling where the overlay moves above
the viewport leading to the issue described in #11152.

By using fixed positioning on the container and the overlay, they are
positioned in the same way and the overlay covers the page content as
expected.

Fixes https://github.com/Automattic/jetpack/issues/11152