Jetpack_Gutenberg: Decrease abstraction when registering extensions
Deprecates the `jetpack_register_block` and `jetpack_register_plugin`
utilities in favor of the new `jetpack_register_gutenberg_extension`.

This new function doesn't actually register an extension, it just
keeps track of the extension that are registered if they are
whitelisted.

This means that the blocks need to be now explicitly registered by the
modules using `register_block_type`. Thanks to this, we remove some
abstraction and allow to build more complex extensions (i.e.:
overriding an existing block).