Sync: Consider whitelisted option filters (#11094)
Currently, when we check whether an option is whitelisted for syncing, we check against the default list of options. This means that any additional options that we register by using the filters aren't considered. This PR changes this, so now the filters will be considered and any options that have been added by using those filters will be synced.

This PR is a companion to D22745-code

#### Changes proposed in this Pull Request:

* Consider option whitelist filters when checking whether an option is whitelisted.

#### Testing instructions:
* See D22745-code

#### Proposed changelog entry for your changes:

* Consider option whitelist filters when checking whether an option is whitelisted.