Only update new settings and use a temporary file to avoid corruption. (#652)
* Only update new settings and use a temporary file to avoid corruption.
Fixes #648
The configuration file can be corrupted if multiple writers try to
update it at once. This patch updates a temporary file that is renamed
to the original name which should fix this.
It also limits updates to settings that have been updated. This should
decrease updates to the file by a huge amount.

* Minor changes to debug lines to indicate function.