Blocks: load each block from a separate file
As we work to add several new modules, the existing module/blocks.php is updated quite often. It causes
a few issues:

- The file keeps getting longer and longer.
- The longer the file, the longer it takes to skim through it.
- Everyone working on that file ends up having to rebase frequently because we are merging changes to
that file often.
- Keeping the file in sync with WordPress.com can be more work if someone does not commit their matching
WordPress.com diff immediately after merging a Jetpack PR.

Splitting things into different files makes things easier to maintain for everyone.

This commit switches to only loading a block file if:
- Gutenberg is available on the site.
- Gutenpack is enabled on the site.
- the block is available in the block list (index.json)

The directory structure was chosen so in the future, it can be used to host the blocks' source code as
well.

For further discussion about this, see:
pafL3P-io-p2