Revert deprecating jetpack_register_block
We introduced jetpack_register_block_type in 11310. But since we already have a wrapper fucntion that we also deprecated. We end up with 2 functions jetpack_register_block_type and jetpack_register_block that do the same thing.
This PR removed the new wrapper function in favour of the old one.