Fix pre_kses filter replacing iframe and blockquotes for URLs
The code for replacing iframe and blockquote for URLs in terms of pre_kses does not really work, as it's
expecting URL in 2nd matching group, which only exists in the first regex. The second one does not even
have 2nd matching group.

In order to make the code work for both regexes, this changeset updates both regexes to contain
instagram.com URL in 1st matching group.

Originally submitted in D24298