Fixes the CDN asset path filter breaking i18n.
Before we would use a certain filter to allow the CDN to properly load files from the wp-includes folder, and the i18n to properly use the path for loading translations. But this fails when loading textdomain for JavaScript because the process uses the same filter. This change fixes that by only applying the filter for URLs that actually contain wp-includes.