Photon: add paypalobjects.com to the list of banned domains.
Fixes #11371

Paypalobjects.com (used to render Payment buttons by Paypal) relies on the Akamai CDN, and the CDN
creates a redirection when you try to access an image. That trips Photon.

Let's consequently add that domain to the list of domains Photon won't try to serve.
a redirection when trying