Gutenpack: Revert deprecating jetpack_register_block (#11328)
We introduced `jetpack_register_block_type` in #11310. But since we already have a wrapper function that we also deprecated. We end up with 2 functions `jetpack_register_block_type` and `jetpack_register_block` that do the same thing. 
This PR removed the new wrapper function in favour of the old one and ads a check that we don't call jetpack_register_block_type without `jetpack/` which is what we expect. 

#### Changes proposed in this Pull Request:
* remove the newly added jetpack_register_block_type in favour of the old jetpack_register_block_type. 


#### Testing instructions:
* Do the tests still pass? 
* Do the blocks still load in the editor? 
* Do you notice any php errors/notices? 

#### Proposed changelog entry
I don't think this needs a changelog entry