Stats: hide the Stats's widget title SVG in Screen Options.
Fixes #11265

In #10958 I changed the widget title and included a SVG file in there, not thinking that
the widget title was also displayed in the Screen Options tab.

WordPress does not currently offer a filter allowing you to change the widget title in
Screen Options but not inside the widgets, so I opted to hide that extra part of the
widget title with CSS in the Screen Options tab.