Sharing: add alt text to OG tags and Twitter Cards
Follow-up from #11339

When we get an alt text when inspecting images for a post, we display it in OG tags and Twitter Meta
cards, allowing Social Networks and third-parties to build proper previews on their sites using that
information.

It's worth noting that we only add that meta information when there is only one image that will be
returned in the post; we can't add multiple alt texts in OG or Twitter meta cards as third-parties would
not be able to know which alt belongs to which image.