Plugin search hints: maintain array keys after sorting (#11467)
This PR solves an issue when tracking the matched module. Previously, the `usort` function was used and it didn't maintain the array keys that were supposed to be the module slugs. This PR changes it to `uasort` that maintains the index association so we can still use the key later.

To verify this, see the live view for the `jetpack_wpa_plugin_search_match_found` event in Tracks.