Publicize: rely on ES to count words for posts.
Summary:
Fixes 3551-gh-jpop-issues

Publicize counts the words in a full post, and compares that number with the number of words in an automatically generated excerpt, to decide how many words should be sent to each Social Network as a preview of the post.
Until now we've relied on `str_word_count` to count the number the words, and while that works well for latin languages, it is problematic with languages with special characters, like Japanese or Cyrillic.
See https://stackoverflow.com/q/23015600/1045686

`es_api_word_count` offers a good workaround on WordPress.com.

Test Plan:
* Prepare your Jetpack site and your WordPress.com sandbox for testing:
    * Sandbox your JP site with the `JETPACK__SANDBOX_DOMAIN` constant.
    * Apply this patch.
    * Set up Publicize to work synchronously on your sandbox.
    * Link your site to a Facebook Page.
* Publish a new post with the content defined in the JPOP issue above.
* Make sure only the first part of the post is pushed to Facebook.

Reviewers: kraftbj

Reviewed By: kraftbj

Tags: #touches_jetpack_files

Differential Revision: D25267-code

This commit syncs r188663-wpcom.