Check for error count increases.
If there are no errors on the changed lines of the new files, there could
still be more errors in one of the new files than in its corresponding old
file. (E.g., the only change to a file is the deletion of a line that
initializes a variable.)

To guard against this, check the error count in the old files to see if
it has increased. (Note that we should ignore lines in the old files that
have changed in the new files, since we know any errors on those lines
have been fixed.)